refactor(ui-theme): move ThemeProvider into theme module, shim old path
